Fix project change bug
Fix a bug that broke folders when you clicked on project title but did not change projects. This meant the "cleanup" method was called but no new li.file items were ever added to the list, so the mutation observer did not fire and load files. Moved new code into loadAllFiles() and restored it so it can be called when a new project is loaded.